requirements.txt
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.2
PyYAML==6.0
scipy==1.10.0
ir_measures @ git+https://github.com/terrierteam/ir_measures.git@9460896